On Mac OS X, use DYLD_LIBRARY_PATH instead of DYLD_FALLBACK_LIBRARY_PATH.
This matches Soylatte's implementation and fixes 'gamma' when using OpenJDK
as a bootstrap VM when building openjdk on Mac OS X. This will result in
OpenJDK's libraries taking precedence; however, this issue already occured
in reverse with DYLD_FALLBACK_LIBRARY_PATH in that Apple's libJPEG.dylib was
preferred over the JDK libjpeg.dylib.

The library search path mechanism will need to be revisited in the future,
and the use of rpath or similar explored.